Various improvements
This update includes:
Camera adjustments such as making it follow the player and zooming it in.

Area change code where when moving into the proper hitbox the scene will transition to a different one (this will be updated to be more professional later).

The players controller has been changed from "Update" to "FixedUpdate" as this will work better with the current movement method.