Minor bugfixes
This update fixes:

The player used to clip into walls, their collision has been placed into a sub object to help avoid this, as the box collider used to rotate with the player.

The players sprite was imported with a bad filter that doesn't work for pixel art making it blurry, this has been fixed.

Added a prefab for the player to the "Objects" folder.

Added collision to signs that were previously missing it.