Basic Interactions
Added the ability to interact with doors and NPCs.

Currently the NPC just states "Talking to NPC!" into the debug log, and the doors just load the scenes for their respective areas.

Next update will optimize this feature and will also either be me fleshing out the NPC interaction or the User Interface.