Established Godot Project
Movement and camera script work as they should and have been commented to hopefully be easy to understand.

I tried to keep the original comments as they were to demonstrate just how similar the two methods are.

The farm and planting codes are in the game but are NOT ready for use.

I will resume progress tomorrow, finalizing some minor details.