Brendan's Farming update, Scene changes, Basics of cutscenes, Basics of stats
Many things, first is Brendan's farming update that was left in the Godot branch.

Time, Health, Stamina, Sanity, and Money are all now stored values. Time is stored in a timer, and the rest are stored in the player. The timer and player send this data to a global script for scene changes.

The biggest change of this update is the addition of code for scene changes. We used to just load a new scene without keeping any data about the player, now spawn points and a global script are utilized to store the player's data and figure out where to spawn them when an area changes.

I also have some setup for cutscenes/dialogue. Right now if you walk up to a NPC and press E it will show a text box with their name.

Next update will flesh out all of these mechanics in various ways and will be added tonight.