Fixed player animation and fishing glitch
Solved the issue where the players sprite was loading and displaying improperly by fixing various import and animation settings.

Added it so that the fishing sprite is shown when fishing begins.

Fixed a glitch with fishing where the player would continue moving if the started fishing while moving. This was fixed by setting the players velocity to zero when fishing begins.